requirements.txt
PyQt5>=5.15.0
PyGithub>=2.0.0
PyMySQL>=1.1.0
PyYAML>=6.0
cryptography>=41.0.0
requests>=2.31.0
paramiko>=3.0.0
pytest>=7.0.0